# Subquery Spin-Off: Reuse results from identical downstream and spun-off queries