--- a/pkg/frontend/querymiddleware/spin_off_subqueries_queryable.go
+++ b/pkg/frontend/querymiddleware/spin_off_subqueries_queryable.go
@@ -6,6 +6,7 @@ import (
 	"context"
 	"fmt"
 	"strings"
+	"sync"
 	"time"
 
 	"github.com/pkg/errors"
@@ -15,6 +16,7 @@ import (
 	"github.com/prometheus/prometheus/util/annotations"
 
 	"github.com/grafana/mimir/pkg/frontend/querymiddleware/astmapper"
+	"github.com/grafana/mimir/pkg/storage/series"
 )
 
 // spinOffSubqueriesQueryable is an implementor of the Queryable interface.
@@ -24,6 +26,12 @@ type spinOffSubqueriesQueryable struct {
 	responseHeaders       *responseHeadersTracker
 	handler               MetricsQueryHandler
 	rangeHandler          MetricsQueryHandler
+
+	queryCacheMu sync.RWMutex
+	queryCache   map[string]storage.SeriesSet
+	// inProgress tracks queries that are currently being executed
+	inProgressMu sync.Mutex
+	inProgress   map[string]chan struct{}
 }
 
 func newSpinOffSubqueriesQueryable(req MetricsQueryRequest, annotationAccumulator *AnnotationAccumulator, next MetricsQueryHandler, rangeHandler MetricsQueryHandler) *spinOffSubqueriesQueryable {
@@ -33,6 +41,8 @@ func newSpinOffSubqueriesQueryable(req MetricsQueryRequest, annotationAccumulato
 		handler:               next,
 		rangeHandler:          rangeHandler,
 		responseHeaders:       newResponseHeadersTracker(),
+		queryCache:            make(map[string]storage.SeriesSet),
+		inProgress:            make(map[string]chan struct{}),
 	}
 }
 
@@ -43,6 +53,10 @@ func (q *spinOffSubqueriesQueryable) Querier(_, _ int64) (storage.Querier, error
 		handler:               q.handler,
 		rangeHandler:          q.rangeHandler,
 		responseHeaders:       q.responseHeaders,
+		queryCache:            q.queryCache,
+		queryCacheMu:          &q.queryCacheMu,
+		inProgress:            q.inProgress,
+		inProgressMu:          &q.inProgressMu,
 	}, nil
 }
 
@@ -60,6 +74,57 @@ type spinOffSubqueriesQuerier struct {
 
 	// Keep track of response headers received when running embedded queries.
 	responseHeaders *responseHeadersTracker
+	queryCache      map[string]storage.SeriesSet
+	queryCacheMu    *sync.RWMutex
+	inProgress      map[string]chan struct{}
+	inProgressMu    *sync.Mutex
+}
+
+// loadFromCacheOrWait loads a query result from cache or waits for an in-progress query to complete.
+// If the query is not in cache and not in progress, it returns nil.
+func (q *spinOffSubqueriesQuerier) loadFromCacheOrWait(ctx context.Context, key string) storage.SeriesSet {
+	// First check if the query is in progress
+	q.inProgressMu.Lock()
+	if waitCh, exists := q.inProgress[key]; exists {
+		q.inProgressMu.Unlock()
+		// Wait for the query to complete
+		select {
+		case <-waitCh:
+			// Query completed, check cache
+		case <-ctx.Done():
+			return storage.ErrSeriesSet(ctx.Err())
+		}
+	} else {
+		// Mark this query as in progress
+		waitCh := make(chan struct{})
+		q.inProgress[key] = waitCh
+		q.inProgressMu.Unlock()
+		return nil
+	}
+
+	q.queryCacheMu.RLock()
+	defer q.queryCacheMu.RUnlock()
+	if set, ok := q.queryCache[key]; ok {
+		concreteSet, ok := set.(*series.ConcreteSeriesSet)
+		if !ok {
+			return storage.ErrSeriesSet(fmt.Errorf("invalid series set type in cache: %T", set))
+		}
+		return concreteSet.ShallowClone()
+	}
+	return nil
+}
+
+// storeInCache stores a query result in the cache and closes the wait channel.
+func (q *spinOffSubqueriesQuerier) storeInCache(key string, set storage.SeriesSet) {
+	q.queryCacheMu.Lock()
+	q.queryCache[key] = set
+	q.queryCacheMu.Unlock()
+
+	q.inProgressMu.Lock()
+	waitCh := q.inProgress[key]
+	delete(q.inProgress, key)
+	q.inProgressMu.Unlock()
+	close(waitCh)
 }
 
 func (q *spinOffSubqueriesQuerier) Select(ctx context.Context, _ bool, hints *storage.SelectHints, matchers ...*labels.Matcher) storage.SeriesSet {
@@ -78,6 +143,11 @@ func (q *spinOffSubqueriesQuerier) Select(ctx context.Context, _ bool, hints *st
 		query, ok := values[astmapper.DownstreamQueryLabelName]
 		if !ok {
 			return storage.ErrSeriesSet(errors.New("missing required labels for downstream query"))
+		}
+
+		set := q.loadFromCacheOrWait(ctx, query)
+		if set != nil {
+			return set
 		}
 
 		downstreamReq, err := q.req.WithQuery(query)
@@ -101,7 +171,11 @@ func (q *spinOffSubqueriesQuerier) Select(ctx context.Context, _ bool, hints *st
 		q.responseHeaders.mergeHeaders(promRes.Headers)
 		q.annotationAccumulator.addInfos(promRes.Infos)
 		q.annotationAccumulator.addWarnings(promRes.Warnings)
-		return newSeriesSetFromEmbeddedQueriesResults([][]SampleStream{resStreams}, hints)
+
+		set = newSeriesSetFromEmbeddedQueriesResults([][]SampleStream{resStreams}, hints, false)
+		q.storeInCache(query, set)
+		return set
+
 	case astmapper.SubqueryMetricName:
 		expr := values[astmapper.SubqueryQueryLabelName]
 		rangeStr := values[astmapper.SubqueryRangeLabelName]
@@ -109,6 +183,14 @@ func (q *spinOffSubqueriesQuerier) Select(ctx context.Context, _ bool, hints *st
 		offsetStr := values[astmapper.SubqueryOffsetLabelName]
 		if expr == "" || rangeStr == "" || stepStr == "" {
 			return storage.ErrSeriesSet(errors.New("missing required labels for subquery"))
+		}
+
+		key := fmt.Sprintf("%s:%s:%s:%s", expr, rangeStr, stepStr, offsetStr)
+
+		var set storage.SeriesSet
+		set = q.loadFromCacheOrWait(ctx, key)
+		if set != nil {
+			return set
 		}
 
 		queryExpr, err := parser.ParseExpr(expr)
@@ -173,7 +255,7 @@ func (q *spinOffSubqueriesQuerier) Select(ctx context.Context, _ bool, hints *st
 			rangeStart = rangeEnd // Move the start to the end of the previous range.
 		}
 
-		sets := make([]storage.SeriesSet, len(rangeQueries))
+		streams := make([][]SampleStream, len(rangeQueries))
 		for idx, req := range rangeQueries {
 			resp, err := q.rangeHandler.Do(ctx, req)
 			if err != nil {
@@ -187,12 +269,15 @@ func (q *spinOffSubqueriesQuerier) Select(ctx context.Context, _ bool, hints *st
 			if err != nil {
 				return storage.ErrSeriesSet(err)
 			}
-			sets[idx] = newSeriesSetFromEmbeddedQueriesResults([][]SampleStream{resStreams}, hints)
+			streams[idx] = resStreams
 			q.annotationAccumulator.addInfos(promRes.Infos)
 			q.annotationAccumulator.addWarnings(promRes.Warnings)
 		}
 
-		return storage.NewMergeSeriesSet(sets, 0, storage.ChainedSeriesMerge)
+		set = newSeriesSetFromEmbeddedQueriesResults(streams, hints, len(streams) > 1)
+		q.storeInCache(key, set)
+		return set
+
 	default:
 		return storage.ErrSeriesSet(errors.Errorf("invalid metric name for the spin-off middleware: %s", name))
 	}